+++ src/sndlink.rs
@@ -1,7 +1,7 @@
 use std::collections::VecDeque;
-use std::future::Future;
+use std::{future::Future, mem, pin::Pin, task::Context, task::Poll};
 
-use ntex::channel::{condition, oneshot};
+use ntex::channel::{condition, oneshot, pool};
 use ntex::util::{BufMut, ByteString, Bytes, BytesMut, Either, Ready};
 use ntex_amqp_codec::protocol::{
     Attach, DeliveryNumber, DeliveryState, Disposition, Error, Flow, MessageFormat,
@@ -13,11 +13,57 @@ use ntex_amqp_codec::Encode;
 use crate::cell::Cell;
 use crate::error::AmqpProtocolError;
 use crate::session::{Session, SessionInner, TransferState};
-use crate::{Delivery, Handle};
+use crate::Handle;
 
 #[derive(Clone)]
 pub struct SenderLink {
     pub(crate) inner: Cell<SenderLinkInner>,
+}
+
+#[derive(Debug)]
+pub(crate) struct DeliveryPromise(
+    Either<pool::Sender<Result<Disposition, AmqpProtocolError>>, (Cell<SenderLinkInner>, Bytes)>,
+);
+
+impl DeliveryPromise {
+    fn new(tx: pool::Sender<Result<Disposition, AmqpProtocolError>>) -> Self {
+        DeliveryPromise(Either::Left(tx))
+    }
+
+    fn new_link(link: Cell<SenderLinkInner>, tag: Bytes) -> Self {
+        DeliveryPromise(Either::Right((link, tag)))
+    }
+
+    pub(crate) fn ready(self, result: Result<Disposition, AmqpProtocolError>) {
+        match self.0 {
+            Either::Left(tx) => {
+                let _r = tx.send(result);
+            }
+            Either::Right((inner, tag)) => (*inner.get_ref().on_disposition)(tag, result),
+        }
+    }
+}
+
+pub(crate) struct SenderLinkInner {
+    pub(crate) id: usize,
+    name: ByteString,
+    session: Session,
+    remote_handle: Handle,
+    delivery_count: SequenceNo,
+    delivery_tag: u32,
+    link_credit: u32,
+    pending_transfers: VecDeque<PendingTransfer>,
+    error: Option<AmqpProtocolError>,
+    closed: bool,
+    on_close: condition::Condition,
+    on_disposition: Box<dyn Fn(Bytes, Result<Disposition, AmqpProtocolError>)>,
+}
+
+struct PendingTransfer {
+    body: Option<TransferBody>,
+    state: TransferState,
+    settle: Option<bool>,
+    message_format: Option<MessageFormat>,
 }
 
 impl std::fmt::Debug for SenderLink {
@@ -34,27 +80,6 @@ impl std::fmt::Debug for SenderLinkInner {
             .field(&std::ops::Deref::deref(&self.name))
             .finish()
     }
-}
-
-pub(crate) struct SenderLinkInner {
-    pub(crate) id: usize,
-    name: ByteString,
-    session: Session,
-    remote_handle: Handle,
-    delivery_count: SequenceNo,
-    delivery_tag: u32,
-    link_credit: u32,
-    pending_transfers: VecDeque<PendingTransfer>,
-    error: Option<AmqpProtocolError>,
-    closed: bool,
-    on_close: condition::Condition,
-}
-
-struct PendingTransfer {
-    body: Option<TransferBody>,
-    state: TransferState,
-    settle: Option<bool>,
-    message_format: Option<MessageFormat>,
 }
 
 impl SenderLink {
@@ -100,6 +125,24 @@ impl SenderLink {
         self.inner.get_mut().send(body, Some(tag))
     }
 
+    pub fn send_no_block<T>(&self, body: T) -> Result<Bytes, AmqpProtocolError>
+    where
+        T: Into<TransferBody>,
+    {
+        self.inner
+            .get_mut()
+            .send_no_block(body, None, self.inner.clone())
+    }
+
+    pub fn send_no_block_with_tag<T>(&self, body: T, tag: Bytes) -> Result<Bytes, AmqpProtocolError>
+    where
+        T: Into<TransferBody>,
+    {
+        self.inner
+            .get_mut()
+            .send_no_block(body, Some(tag), self.inner.clone())
+    }
+
     pub fn settle_message(&self, id: DeliveryNumber, state: DeliveryState) {
         self.inner.get_mut().settle_message(id, state)
     }
@@ -120,6 +163,13 @@ impl SenderLink {
 
     pub fn on_close(&self) -> condition::Waiter {
         self.inner.get_ref().on_close.wait()
+    }
+
+    pub fn on_disposition<F>(&self, f: F)
+    where
+        F: Fn(Bytes, Result<Disposition, AmqpProtocolError>) + 'static,
+    {
+        self.inner.get_mut().on_disposition = Box::new(f);
     }
 }
 
@@ -143,6 +193,7 @@ impl SenderLinkInner {
             closed: false,
             delivery_tag: 0,
             on_close: condition::Condition::new(),
+            on_disposition: Box::new(|_, _| ()),
         }
     }
 
@@ -167,6 +218,7 @@ impl SenderLinkInner {
             error: None,
             closed: false,
             on_close: condition::Condition::new(),
+            on_disposition: Box::new(|_, _| ()),
         }
     }
 
@@ -188,7 +240,7 @@ impl SenderLinkInner {
         // drop pending transfers
         for tr in self.pending_transfers.drain(..) {
             if let TransferState::First(tx, _) | TransferState::Only(tx, _) = tr.state {
-                let _ = tx.send(Err(err.clone()));
+                tx.ready(Err(err.clone()));
             }
         }
 
@@ -260,13 +312,9 @@ impl SenderLinkInner {
                 }
             }
         }
-
-        if flow.echo() {
-            // todo: send flow
-        }
     }
 
-    pub(crate) fn send<T: Into<TransferBody>>(&mut self, body: T, tag: Option<Bytes>) -> Delivery {
+    fn send<T: Into<TransferBody>>(&mut self, body: T, tag: Option<Bytes>) -> Delivery {
         if let Some(ref err) = self.error {
             Delivery::Resolved(Err(err.clone()))
         } else {
@@ -299,8 +347,8 @@ impl SenderLinkInner {
 
                 self.send_inner(
                     chunk.into(),
-                    TransferState::First(delivery_tx, tag),
                     message_format,
+                    TransferState::First(DeliveryPromise::new(delivery_tx), tag),
                 );
 
                 loop {
@@ -308,26 +356,89 @@ impl SenderLinkInner {
 
                     // last chunk
                     if body.is_empty() {
-                        self.send_inner(chunk.into(), TransferState::Last, message_format);
+                        self.send_inner(chunk.into(), message_format, TransferState::Last);
                         break;
                     }
 
-                    self.send_inner(chunk.into(), TransferState::Continue, message_format);
+                    self.send_inner(chunk.into(), message_format, TransferState::Continue);
                 }
             } else {
-                let state = TransferState::Only(delivery_tx, self.get_tag(tag));
-                self.send_inner(body, state, message_format);
+                let st = TransferState::Only(DeliveryPromise::new(delivery_tx), self.get_tag(tag));
+                self.send_inner(body, message_format, st);
             }
 
             Delivery::Pending(delivery_rx)
         }
     }
 
+    fn send_no_block<T: Into<TransferBody>>(
+        &mut self,
+        body: T,
+        tag: Option<Bytes>,
+        link: Cell<SenderLinkInner>,
+    ) -> Result<Bytes, AmqpProtocolError> {
+        if let Some(ref err) = self.error {
+            Err(err.clone())
+        } else {
+            let body = body.into();
+            let message_format = body.message_format();
+
+            let max_frame_size = self.session.inner.get_ref().max_frame_size();
+            let max_frame_size = if max_frame_size > 2048 {
+                max_frame_size - 2048
+            } else if max_frame_size == 0 {
+                usize::MAX
+            } else {
+                max_frame_size
+            };
+
+            // body is larger than allowed frame size, send body as a set of transfers
+            if body.len() > max_frame_size {
+                let mut body = match body {
+                    TransferBody::Data(data) => data,
+                    TransferBody::Message(msg) => {
+                        let mut buf = BytesMut::with_capacity(msg.encoded_size());
+                        msg.encode(&mut buf);
+                        buf.freeze()
+                    }
+                };
+
+                let chunk = body.split_to(std::cmp::min(max_frame_size, body.len()));
+                let tag = self.get_tag(tag);
+
+                self.send_inner(
+                    chunk.into(),
+                    message_format,
+                    TransferState::First(DeliveryPromise::new_link(link, tag.clone()), tag.clone()),
+                );
+
+                loop {
+                    let chunk = body.split_to(std::cmp::min(max_frame_size, body.len()));
+
+                    // last chunk
+                    if body.is_empty() {
+                        self.send_inner(chunk.into(), message_format, TransferState::Last);
+                        break;
+                    }
+
+                    self.send_inner(chunk.into(), message_format, TransferState::Continue);
+                }
+                Ok(tag)
+            } else {
+                let tag = self.get_tag(tag);
+                let st =
+                    TransferState::Only(DeliveryPromise::new_link(link, tag.clone()), tag.clone());
+                self.send_inner(body, message_format, st);
+                Ok(tag)
+            }
+        }
+    }
+
     fn send_inner(
         &mut self,
         body: TransferBody,
-        state: TransferState,
         message_format: Option<MessageFormat>,
+        state: TransferState,
     ) {
         if self.link_credit == 0 {
             log::trace!(
@@ -441,18 +552,36 @@ impl SenderLinkBuilder {
             Err(_) => Err(AmqpProtocolError::Disconnected),
         }
     }
+}
 
-    pub async fn attach_simple(self) -> Result<SenderLink, AmqpProtocolError> {
-        let result = self
-            .session
-            .get_mut()
-            .attach_local_sender_link(self.frame)
-            .await;
+enum Delivery {
+    Resolved(Result<Disposition, AmqpProtocolError>),
+    Pending(pool::Receiver<Result<Disposition, AmqpProtocolError>>),
+    Gone,
+}
 
-        match result {
-            Ok(Ok(inner)) => Ok(SenderLink { inner }),
-            Ok(Err(e)) => Err(e),
-            Err(_) => Err(AmqpProtocolError::Disconnected),
+impl Future for Delivery {
+    type Output = Result<Disposition, AmqpProtocolError>;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        if let Delivery::Pending(ref mut receiver) = *self {
+            return match Pin::new(receiver).poll(cx) {
+                Poll::Ready(Ok(r)) => Poll::Ready(r),
+                Poll::Pending => Poll::Pending,
+                Poll::Ready(Err(e)) => {
+                    trace!("delivery oneshot is gone: {:?}", e);
+                    Poll::Ready(Err(AmqpProtocolError::Disconnected))
+                }
+            };
         }
+
+        let old_v = mem::replace(&mut *self, Delivery::Gone);
+        if let Delivery::Resolved(r) = old_v {
+            return match r {
+                Ok(state) => Poll::Ready(Ok(state)),
+                Err(e) => Poll::Ready(Err(e)),
+            };
+        }
+        panic!("Polling Delivery after it was polled as ready is an error.");
     }
 }