+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [0.5.0-b.3] - 2021-08-xx
 
+* Add stream handling for transfer dispositions
+
 * Refactor sender link disposition handling
 
 ## [codec-0.6.1] - 2021-08-0x